backend/requirements.txt
fastapi
fastapi-cors
uvicorn
python-multipart
google-cloud-documentai
google-cloud-firestore
google-cloud-bigquery
google-cloud-storage
pydantic
google-cloud-aiplatform
